Refactor Human AI controller access through contexts

+++ code/modules/mob/living/carbon/human/ai/action_datums/reload.dm
@@ -4,7 +4,11 @@
 	required_ai_modules = list(/datum/human_ai_module/guns, /datum/human_ai_module/inventory)
 	var/currently_reloading
 
-/datum/ai_action/reload/get_weight(datum/human_ai_brain/brain)
+/datum/ai_action/reload/get_context_weight(datum/human_ai_context/context)
+	var/datum/human_ai_brain/brain = context?.brain
+	if(!brain)
+		return 0
+
 	if(brain.has_tried_reload())
 		return 0
 
@@ -25,6 +29,10 @@
 	if(. == ONGOING_ACTION_COMPLETED)
 		return .
 
+	var/datum/human_ai_brain/brain = context?.brain
+	if(!brain)
+		return ONGOING_ACTION_COMPLETED
+
 	if(currently_reloading)
 		return ONGOING_ACTION_UNFINISHED
 
@@ -38,10 +46,15 @@
 /datum/ai_action/reload/proc/reload()
 	set waitfor = FALSE
 
+	var/datum/human_ai_brain/brain = context?.brain
+	var/datum/human_tied_controller/controller = context?.controller
+	if(!brain || !controller)
+		return
+
 	var/obj/item/weapon/gun/primary_weapon = brain.get_primary_weapon()
 	var/datum/human_ai_firearm_profile/gun_data = brain.get_gun_data()
 	if(gun_data.disposable)
-		brain.tied_controller.drop_held_item(primary_weapon)
+		controller.drop_held_item(primary_weapon)
 		brain.unqueue_pickup(primary_weapon)
 		brain.set_primary_weapon(null)
 		qdel(src)
@@ -50,27 +63,34 @@
 	currently_reloading = TRUE
 
 	/// Find ammo
-	var/datum/human_ai_firearm_context/context = new(primary_weapon, brain)
-	var/datum/human_ai_firearm_handler/handler = context.get_handler()
-	var/obj/item/reload_item = handler?.find_reload_item(context)
+	var/datum/human_ai_firearm_context/firearm_context = new(primary_weapon, brain)
+	var/datum/human_ai_firearm_handler/handler = firearm_context.get_handler()
+	var/obj/item/reload_item = handler?.find_reload_item(firearm_context)
 	if(!reload_item)
-		qdel(context)
+		qdel(firearm_context)
 		brain.mark_tried_reload()
 		qdel(src)
 		return
 
-	context.set_reload_item(reload_item)
+	firearm_context.set_reload_item(reload_item)
 	brain.say_reload_line()
-	handler.do_reload(context)
-	qdel(context)
+	handler.do_reload(firearm_context)
+	qdel(firearm_context)
+
+	if(!context?.can_continue())
+		return
 
 	/// When do_reload() stops sleeping, let us check things one last time
 	currently_reloading = FALSE
 
 /datum/ai_action/reload/proc/primary_ammo_search()
+	var/datum/human_ai_brain/brain = context?.brain
+	if(!brain)
+		return
+
 	var/obj/item/weapon/gun/primary_weapon = brain.get_primary_weapon()
-	var/datum/human_ai_firearm_context/context = new(primary_weapon, brain)
-	var/datum/human_ai_firearm_handler/handler = context.get_handler()
-	var/obj/item/reload_item = handler?.find_reload_item(context)
-	qdel(context)
+	var/datum/human_ai_firearm_context/firearm_context = new(primary_weapon, brain)
+	var/datum/human_ai_firearm_handler/handler = firearm_context.get_handler()
+	var/obj/item/reload_item = handler?.find_reload_item(firearm_context)
+	qdel(firearm_context)
 	return reload_item
